todo_app/requirements.txt
asgiref==3.5.0
Django==3.2.12
djangorestframework==3.13.1
importlib-metadata==4.11.1
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0
